% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ftrend.R
\name{Ftrend}
\alias{Ftrend}
\title{Generate a trend in fishing mortality}
\usage{
Ftrend(yr.st, yr.end, curF, F.pat = c("stable", "inc", "dec"), Fcv = 0.3,
  plot = TRUE)
}
\arguments{
\item{yr.st}{Numeric. First year - unfished}

\item{yr.end}{Numeric. Current year - F = `curF`}

\item{curF}{Numeric. Current apical fishing mortality}

\item{F.pat}{Character. F-pattern. Choices are: 'stable', 'inc' (increasing),
and 'dec' (decreasing)}

\item{plot}{Logical. Should the F-trend be plotted?}

\item{Numeric.}{Fcv inter-annual CV of F}
}
\value{
A numeric vector of fishing mortality
}
\description{
Generate a trend in fishing mortality
}
\examples{
Ftrend(1950, 1990, 0.1, 'stable', 0.3, TRUE)
}
