% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData.R
\name{MakeFleet}
\alias{MakeFleet}
\title{Make a Fleet List with default values}
\usage{
MakeFleet()
}
\value{
A list of Fleet Values
}
\description{
Make a Fleet List with default values
}
\examples{
MakeFleet()
}
