% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData.R
\name{MakeStock}
\alias{MakeStock}
\title{Make a Stock List with default values}
\usage{
MakeStock()
}
\value{
A list of Stock Values
}
\description{
Make a Stock List with default values
}
\examples{
MakeStock()
}
