% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData.R
\name{MakeSurvey}
\alias{MakeSurvey}
\title{Make a Survey List with default values}
\usage{
MakeSurvey()
}
\value{
A list of Survey Values
}
\description{
Make a Survey List with default values
}
\examples{
MakeSurvey()
}
