% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimData.R
\name{SimData}
\alias{SimData}
\title{Simulate Data}
\usage{
SimData(Stock = MakeStock(), Fleet = MakeFleet(), Survey = MakeSurvey(),
  Control = list(ngtg = 101), plot = FALSE)
}
\arguments{
\item{Stock}{A Stock List}

\item{Fleet}{A Fleet List}

\item{Survey}{A Survey List}

\item{Control}{A named list of control options}

\item{plot}{Logical. Show the plots?}
}
\value{
A data.frame of sampled age, length, weight, and maturity data
}
\description{
Simulate Data
}
\examples{
\dontrun{
MyData <- SimData()
}
}
